Create main.py
This is a simple python file for writing code for programming contests. The environment variable 'KANGOONIE' redirects the standard IO to local files when the variable is defined. Otherwise all IO operations are done using standard IO (i.e console)
To get it to work you will need to define your environment variable, and create the files in.txt and out.txt in the same directory level as this file.
Enjoy!